Create activatesshkey package
